--- src/components/App/index.ts
@@ -1,3 +1,0 @@
-import App from './App'
-
-export default App
